+++ src/components/Footer.native.tsx
@@ -1,24 +1,26 @@
 import React from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import { useDispatch } from 'react-redux'
 import * as pkg from '../../package.json'
 // import { TUTORIAL2_STEP_SUCCESS } from '../constants'
 import { logout, showModal } from '../action-creators'
 // import { getSetting, isTutorial } from '../selectors'
 import { scaleFontDown, scaleFontUp } from '../action-creators/scaleSize'
-import { State } from '../util/initialState'
+// import { State } from '../util/initialState'
 import { View, StyleSheet, TouchableOpacity } from 'react-native'
 import { Text } from './Text.native'
+import { useFooterUseSelectors } from '../hooks/Footer.useSelectors'
 
 /** A footer component with some useful links. */
 const Footer = () => {
   const dispatch = useDispatch()
-  const authenticated = useSelector((state: State) => state.authenticated)
-  const user = useSelector((state: State) => state.user)
-  const status = useSelector((state: State) => state.status)
-  // const tutorialStep = useSelector((state: State) => +(getSetting(state, 'Tutorial Step') || 1))
-  // const isPushing = useSelector((state: State) => state.isPushing)
-  // const isTutorialOn = useSelector(isTutorial)
-  const pushQueueLength = useSelector((state: State) => state.pushQueue.length)
+  const { authenticated, user, status, pushQueueLength } = useFooterUseSelectors()
+  // const authenticated = useSelector((state: State) => state.authenticated)
+  // const user = useSelector((state: State) => state.user)
+  // const status = useSelector((state: State) => state.status)
+  // // const tutorialStep = useSelector((state: State) => +(getSetting(state, 'Tutorial Step') || 1))
+  // // const isPushing = useSelector((state: State) => state.isPushing)
+  // // const isTutorialOn = useSelector(isTutorial)
+  // const pushQueueLength = useSelector((state: State) => state.pushQueue.length)
 
   // if (isTutorialOn && tutorialStep !== TUTORIAL2_STEP_SUCCESS) return null
